+++ Java/rei-dos-piratas/src/main/java/br/com/fiap/rei_dos_piratas/interfaces/dto/ProdutoInDto.java
@@ -7,6 +7,8 @@ import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.NotNull;
 import org.hibernate.validator.constraints.Length;
 
+import java.math.BigDecimal;
+
 public record ProdutoInDto(
         @NotNull(message = "O nome do produto não pode ser nulo")
         @Length(min = 10, max = 150, message = "O nome do produto deve ter entre 10 e 150 caracteres")
@@ -21,7 +23,7 @@ public record ProdutoInDto(
         
         @Digits(fraction = 2, integer = 6, message = "O preço do produto deve ter até 8 digitos com 2 dígitos após a vírgula")
         @DecimalMin(value = "0.0", inclusive = false, message = "O preço não pode ser negativo")
-        float preco,
+        BigDecimal preco,
         
         @Digits(fraction = 0, integer = 6, message = "O estoque deve ter 6 dígitos sem dígitos após a vírgula")
         @DecimalMin(value = "0", inclusive = false, message = "O estoque não pode ser negativo")
@@ -29,15 +31,15 @@ public record ProdutoInDto(
         
         @Digits(fraction = 0, integer = 6, message = "A altura do produto deve ter até 8 digitos com dois após a vírgula")
         @DecimalMin(value = "0", inclusive = false, message = "A altura do produto não pode ser negativa")
-        float altura,
+        BigDecimal altura,
         
         @Digits(fraction = 0, integer = 6, message = "A largura do produto deve ter até 8 digitos com dois após a vírgula")
         @DecimalMin(value = "0", inclusive = false, message = "A largura do produto não pode ser negativa")
-        float largura,
+        BigDecimal largura,
         
         @Digits(fraction = 0, integer = 6, message = "A profundidade do produto deve ter até 8 digitos com dois após a vírgula")
         @DecimalMin(value = "0", inclusive = false, message = "A profundidade do produto não pode ser negativa")
-        float profundidade,
+        BigDecimal profundidade,
 
         @NotNull(message = "A condição do produto é obrigatória(NOVO ou USADO)")
         CondicaoEnum condicao,
